V backend: specify type in variable initializer.
The code without exact types in variable initializers used to compile
in older versions of V, but not anymore.

+++ examples/v/fill/02_fill.v
@@ -55,7 +55,7 @@ loop:
     yyrecord.token = yyrecord.yycursor
     
 //line "v/fill/02_fill.v":58
-    mut yych := 0
+    mut yych := u8(0)
     if yyrecord.yylimit <= yyrecord.yycursor {
         r := fill(mut yyrecord, 1); if r != 0 { return r }
     }
